Burnt Out
Yeah, this burnt me out. Big Time. so i decided not to pursue the use of
a webcam to capture images, because there was a problem with using teh
webcam, and using a webcam to talk to kivy is not relevant with our
mission. So then I decided to work on a project tu=o use the image
widget. This burnt me out. We recieve an image, an it is in our memory.
The dirty way to deal with this is to save tan image onto a file, then
load that image onto kivy. and refresh every couple of moments,
depending on desired framerate. However, the better way to do this is to
directly load the image onto kivy from memory, never saving the image
onto the computer I tried. I tried very much. I'm done, for now. It was
not a complete waste. I learnt about the very useful stringio library
that come with python. This allows us to create a file in the memory. so
that when we recieve the iage, rather than "downloading" the image onto
the computer, we place the image into a file in the memory, then simple
refer to that image. The problem was that kivy is too high level. The
comand to load an image uses a source option, where we tell kivy where
the picture exists. This source command only accepts strings that lead
to a file in the hard drive to look for an image, perfect for the first
dirty way of doing this. So, I had to dig deeper into kivy. Learn how
the source function works, so that I call function deeper than Image.
Unfortunately, it became impractical for now, considering this is only
for testing, and time is essential, where a much easier, though less
desirable, but surer to work method exists. But for now, I am just
pushing the changes up to I stopped trying to work with stringio.