2 comps
after much blood, sweat, and tears, i got crossbar to work on 2
computers!!!!!!! (aparently, local means computer-you-are-using, not
network-you-are-using. I was running the scripts on both computers on
the localhost's network, which means, i was isolating all events to that
computer. Basically, i was limiting my script to only look for events,
coming from the local host, or that computer itself. After a couple a
hours, i got it. simply replace the application runner url. So search
for events from the router (symbolized and identified by the ip address)
instead of from the localhost. (before i was putting the ip address to
search for to be "localhost". but yeah. this script should work on any
computer, (such as a pi, wink wink).